requirements.txt
matplotlib==3.5.1
numpy==1.19.2
pandas==1.4.2
pypianoroll==1.0.4
PyYAML==6.0
seaborn==0.11.2
tensorflow==2.8.0
tensorflow_gpu==2.4.0
